requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
openai>=1.3.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
duckduckgo-search>=4.0.0
requests>=2.28.0

